demo_script.py
#!/usr/bin/env python3
"""
Demo script to showcase CyberTool functionality
"""

import subprocess
import time
import sys

def run_demo():
    """Run a quick demo of CyberTool"""
    print("🚀 CyberTool - All-in-One Cybersecurity Toolkit")
    print("=" * 50)
    print("\n📋 Available Tools:")
    print("1. 🔍 Subdomain Enumeration")
    print("2. 🔒 PDF Protection Tool") 
    print("3. 🔓 PDF Cracker")
    print("4. 🌐 Network Scanner")
    print("5. 🚪 Port Scanner")
    print("6. 💣 Password Cracker")
    print("7. 🔐 SSH Cracker")
    print("8. 📁 FTP Cracker")
    print("9. ℹ️  Information Stealer")
    
    print("\n✅ Installation completed successfully!")
    print("📁 Project structure created:")
    print("   ├── main.py (Main CLI interface)")
    print("   ├── tools/ (Individual tool modules)")
    print("   ├── wordlists/ (Common passwords/usernames)")
    print("   ├── examples/ (Usage examples)")
    print("   ├── requirements.txt (Dependencies)")
    print("   └── README.md (Documentation)")
    
    print("\n🎯 To start using CyberTool:")
    print("   python main.py")
    
    print("\n⚠️  IMPORTANT REMINDER:")
    print("   Use only on systems you own or have permission to test!")
    print("   This tool is for educational and ethical purposes only.")
    
    print("\n📚 For detailed usage examples, see:")
    print("   examples/basic_usage.md")
    print("   README.md")
    
    print("\n🔧 All dependencies installed:")
    try:
        import colorama
        import pyfiglet
        import tqdm
        import requests
        import paramiko
        from Crypto.Cipher import AES
        import dns
        import nmap
        import PyPDF2
        import psutil
        print("   ✅ All required packages installed!")
    except ImportError as e:
        print(f"   ❌ Missing package: {e}")
        print("   Run: pip install -r requirements.txt")
    
    print("\n🌟 CyberTool is ready to use!")

if __name__ == "__main__":
    run_demo()