requirements.txt
requests==2.31.0
python-nmap==0.7.1
pycryptodome==3.19.0
paramiko==3.3.1
colorama==0.4.6
tqdm==4.66.1
pyfiglet==0.8.post1
click==8.1.7
dnspython==2.4.2
PyPDF2==3.0.1
pillow==10.1.0
# Standard library modules (no need to install)
# hashlib
# socket
# threading
# concurrent.futures